app/requirements.txt
fastmcp>=0.9.0
aiohttp>=3.9.0
python-dotenv>=1.0.0